+++ src/base/b-list/interface.ts
@@ -6,44 +6,88 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+import { HintPosition } from 'global/g-hint/interface';
+
 export interface Item {
+	/**
+	 * Item label text
+	 */
 	label?: string;
+
+	/**
+	 * Item value
+	 */
 	value?: unknown;
+
+	/**
+	 * If this option is provided the component will generate a link for this item
+	 */
 	href?: string;
-	info?: string;
+
+	/**
+	 * True if the item is active
+	 */
 	active?: boolean;
+
+	/**
+	 * True if the item is hidden
+	 */
 	hidden?: boolean;
+
+	/**
+	 * True if the item is in-progress
+	 */
 	progress?: boolean;
+
+	/**
+	 * Exterior modifier of the item
+	 */
 	exterior?: string;
-	classes?: Dictionary<string>;
+
+	/**
+	 * Tooltip text of the item
+	 */
 	hint?: string;
 
+	/**
+	 * Tooltip position to show
+	 */
+	hintPos?: HintPosition;
+
+	/**
+	 * Icon to show before a label
+	 */
 	preIcon?: string;
 
 	/**
-	 * Name of used component to show `preIcon`
+	 * Name of the used component to show `preIcon`
 	 * @default `'b-icon'`
 	 */
 	preIconComponent?: string;
 
 	/**
-	 * Text of a hint that is shown on hovering of `preIcon`
+	 * Icon to show after a label
 	 */
-	preIconHint?: string;
+	icon?: string;
 
 	/**
-	 * Position of a `preIcon` hint to show
+	 * Name of the used component to show `icon`
+	 * @default `'b-icon'`
 	 */
-	preIconHintPos?: string;
-
-	icon?: string;
 	iconComponent?: string;
-	iconHint?: string;
-	iconHintPos?: string;
+
+	/**
+	 * Component to show "in-progress" state of the item
+	 */
 	progressIcon?: string;
 
 	/**
-	 * Dictionary with extra HTML attributes of an item
+	 * Map of additional classes of the item
+	 */
+	classes?: Dictionary<string>;
+
+	/**
+	 * Map of extra attributes of the item
 	 */
 	attrs?: Dictionary;
 }