+++ src/base/b-list/CHANGELOG.md
@@ -19,6 +19,10 @@ Changelog
 * Renamed `removeActive` -> `unsetActive`
 * Renamed `normalizeOptions` -> `normalizeItems`
 * Now the `active` getter returns a Set with `multiple = true`
+* Removed `block info` from the template
+* Renamed `block text` to `block value`
+* Renamed `&__link-text` to `&__link-value`
+* Deprecated `&__el`
 
 #### :bug: Bug Fix